MFRC522/requirements.txt
RPi.GPIO==0.6.3
spidev